ugc/etl/requirements.txt
kafka-python==2.0.2 
clickhouse-driver==0.2.7
backoff==2.2.1